Fixed unique username and email

+++ src/routes/user.js
@@ -3,9 +3,12 @@ const User = require('../models/user')
 const router = new express.Router()
 const auth  = require('../middleware/auth')
 
-
+// Get all users
 router.get('/users/getAll', async(req,res)=>{
     try{
+        if(req.body.key!==process.env.key){
+            return res.sendStatus(404)
+        }
         const users = await User.find();
         res.status(201).send(users)
     }catch(e){
@@ -13,8 +16,13 @@ router.get('/users/getAll', async(req,res)=>{
     }
     
 })
+
+// Dump all users
 router.delete('/users/dumpAll', async(req,res)=>{
     try{
+        if(req.body.key!==process.env.key){
+            return res.sendStatus(404)
+        }
          await User.deleteMany();
         res.status(201).send({"message":"Users Deleted"})
     }catch(e){
@@ -22,32 +30,54 @@ router.delete('/users/dumpAll', async(req,res)=>{
     }
     
 })
+
+
+// --------------------------Auth Routes--------------------------------
+// Signup user
 router.post('/users/signup',async (req, res)=>{
-    const user =  new User(req.body)
+        const user =  new User(req.body)
         try{
             await user.save()
             const token = await user.generateToken()
-            console.log(`token in routes ${token}`);
             res.status(201).send({user,token})
+
         }
         catch(e){
-            console.log(e);
-           res.status(400).send(e)    
+            // Catch what kind of error thrown
+            if(e.errors){
+                let error = e.errors
 
+                // Send error if duplicate username or email
+                if(error.username){
+                    res.status(400).send(error.username.message)    
+                }
+                if(error.email){
+                    res.status(400).send(error.email.message)    
+                }
+            }
+            res.status(400).send(e)    
         }
+
 })
 
+
+// Login user
 router.post('/users/login', async(req,res)=>{
     try {
         const user = await User.findByCredentials(req.body.email, req.body.password)
+
         const token = await user.generateToken()
+        
         res.send({user,token} )
     } catch (error) {
         console.log(error);
-        res.status(400).send()
+        res.status(400).send(error)
     }
-}) 
-router.post('/users/logout',auth, async(req,res)=>{
+})
+
+
+// Logout User
+router.patch('/users/logout',auth, async(req,res)=>{
     try {
         req.user.tokens = req.user.tokens.filter((token) => {
           return token.token !== req.token;
@@ -60,6 +90,20 @@ router.post('/users/logout',auth, async(req,res)=>{
       }
 }) 
 
+// Logout User from everywhere
+router.patch('/users/logoutAll',auth, async(req,res)=>{
+    try {
+        req.user.tokens =[]
+        await req.user.save();
+        res.sendStatus(200);
+      } catch (error) {
+        console.log(error);
+        res.sendStatus(404);
+      }
+})
+// --------------------------Auth Routes ends--------------------------------
+
+
 
 router.get('/users/me', auth, async(req, res)=>{
     try{
@@ -71,5 +115,16 @@ router.get('/users/me', auth, async(req, res)=>{
 })
 
 
+// const checkforUniqueEmailorUsername=async(email,username)=>{
+//     const emailUser = await User.findOne({email})
+
+//     if(emailUser){
+//         return {error: "Email already exists"}
+//     }
+//     const usernameUser = await User.findOne({username})
+//     if(usernameUser){
+//         return {error: "Username already exists"}
+//     }
+// }
 
 module.exports = router
